+++ branches/custom-train-tracks/MTA10/gui/CGUIGridList_Impl.cpp
@@ -239,7 +239,7 @@ CGUIListItem* CGUIGridList_Impl::GetItem ( int iRow, int hColumn )
     try
     {
         CEGUI::MultiColumnList* pMultiColumnList = reinterpret_cast < CEGUI::MultiColumnList* > ( m_pWindow );
-        if ( iRow >= pMultiColumnList->getRowCount() || GetColumnIndex( hColumn ) >= pMultiColumnList->getColumnCount() )
+        if ( (uint)iRow >= pMultiColumnList->getRowCount() || (uint)GetColumnIndex( hColumn ) >= pMultiColumnList->getColumnCount() )
             return NULL;
 
         // Grab the item at the chosen row / column
@@ -263,7 +263,7 @@ const char* CGUIGridList_Impl::GetItemText ( int iRow, int hColumn )
     try
     {
         CEGUI::MultiColumnList* pMultiColumnList = reinterpret_cast < CEGUI::MultiColumnList* > ( m_pWindow );
-        if ( iRow >= pMultiColumnList->getRowCount() || GetColumnIndex( hColumn ) >= pMultiColumnList->getColumnCount() )
+        if ( (uint)iRow >= pMultiColumnList->getRowCount() || (uint)GetColumnIndex( hColumn ) >= pMultiColumnList->getColumnCount() )
             return "";
 
         // Grab the item at the chosen row / column
